Tehtäviä

+++ Tehtävä 5/osa 2.py
@@ -3,7 +3,7 @@ while True:
     luku = input("Anna luku: ")
     if luku == "":
         break
-    lista.append(int(luku))
+    lista.append(float(luku))
 lista.sort(reverse=True)
 for i in range(5):
     print(lista[i])
